Moved ConstructionSite into model namespace.
