+++ service/src/main/java/restwars/model/building/ConstructionSite.java
@@ -1,8 +1,7 @@
-package restwars.service.building;
+package restwars.model.building;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
-import restwars.model.building.BuildingType;
 
 import java.util.UUID;
