+++ storage/src/main/java/restwars/storage/building/JooqConstructionSiteDAO.java
@@ -5,7 +5,7 @@ import org.jooq.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import restwars.model.building.BuildingType;
-import restwars.service.building.ConstructionSite;
+import restwars.model.building.ConstructionSite;
 import restwars.service.building.ConstructionSiteDAO;
 import restwars.service.unitofwork.UnitOfWorkService;
 import restwars.storage.jooq.AbstractJooqDAO;
